feat: add new article on audio data in deep learning
This commit adds a new article on the topic of audio data in deep learning. The article covers the fundamentals of sound and its transformation into digital data, as well as techniques for analyzing and preparing audio data for use in machine learning models. The key changes include:

- Addition of a new HTML file `audio_en_deep_learning.html` that contains the article content
- Addition of a new link to the article in the `index.html` file

The goal of these changes is to provide more comprehensive coverage of the different types of data used in deep learning, beyond just images, which is a common focus.